olddemo/requirements.txt
absl-py==0.9.0
aiomysql==0.0.20
aiosqlite==0.15.0
altgraph==0.17
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
future==0.18.2
gast==0.3.3
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
icmplib==1.1.1
idna==2.9
iso8601==0.1.13
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
pefile==2019.4.18
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
PyInstaller==3.6
pylint==2.5.3
PyMySQL==0.9.2
PyPika==0.40.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
pytz==2020.5
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==1.5
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
tortoise-orm==0.16.16
typing-extensions==3.7.4.3
urllib3==1.25.8
vthread==0.1.4
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
redis
sshtunnel